Refactor Codec Msg
